up bai case study task 4

+++ module4/case_study/furama-resort/src/main/java/com/codegym/model/Role.java
@@ -7,20 +7,20 @@ import java.util.Set;
 public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int roleId;
+    private Long roleId;
     private String roleName;
 
     @ManyToMany(mappedBy = "roles", fetch = FetchType.LAZY, cascade = CascadeType.ALL)
     private Set<User> users;
-    
+
     public Role() {
     }
 
-    public int getRoleId() {
+    public Long getRoleId() {
         return roleId;
     }
 
-    public void setRoleId(int roleId) {
+    public void setRoleId(Long roleId) {
         this.roleId = roleId;
     }
 
@@ -39,5 +39,5 @@ public class Role {
     public void setUsers(Set<User> users) {
         this.users = users;
     }
-    
+
 }
